+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 -   Drawing on FOW layer blocking UI
 -   Badge toggle not working properly
+-   Group leader not properly set serverside on paste
 
 ## [0.22.1] - 2020-08-27