+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 -   Drawing on FOW layer blocking UI
 -   Badge toggle not working properly
 -   Group leader not properly set serverside on paste
+-   Server error on shape paste due to aura type change
 
 ## [0.22.1] - 2020-08-27